Refactor resolvers
Before the resolve code was bound to a budget. Now it is its own package
that can do resolution (getting from a selector to a set of nodes).
Next step will be to refactor the code to fetch and calculate disruptions.
