+++ internal/controller/applicationdisruptionbudget_controller.go
@@ -19,25 +19,19 @@ package controller
 import (
 	"context"
 	"fmt"
-	"io/ioutil"
+	"io"
 	"log"
 	"net/http"
-	"strings"
 
 	"k8s.io/apimachinery/pkg/api/errors"
-	"k8s.io/apimachinery/pkg/fields"
-	"k8s.io/apimachinery/pkg/labels"
 	"k8s.io/apimachinery/pkg/runtime"
-	"k8s.io/apimachinery/pkg/selection"
-	"k8s.io/apimachinery/pkg/types"
 	ctrl "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
 	nodedisruptionv1alpha1 "github.com/criteo/node-disruption-controller/api/v1alpha1"
+	"github.com/criteo/node-disruption-controller/pkg/resolver"
 
 	"github.com/golang-collections/collections/set"
-	corev1 "k8s.io/api/core/v1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
 
 // ApplicationDisruptionBudgetReconciler reconciles a ApplicationDisruptionBudget object
@@ -76,6 +70,7 @@ func (r *ApplicationDisruptionBudgetReconciler) Reconcile(ctx context.Context, r
 	resolver := ApplicationDisruptionBudgetResolver{
 		ApplicationDisruptionBudget: adb,
 		Client:                      r.Client,
+		Resolver:                    resolver.Resolver{Client: r.Client},
 	}
 
 	resolver.Sync(ctx)
@@ -97,11 +92,12 @@ func (r *ApplicationDisruptionBudgetReconciler) SetupWithManager(mgr ctrl.Manage
 type ApplicationDisruptionBudgetResolver struct {
 	ApplicationDisruptionBudget *nodedisruptionv1alpha1.ApplicationDisruptionBudget
 	Client                      client.Client
+	Resolver                    resolver.Resolver
 }
 
 // Sync ensure the budget's status is up to date
 func (r *ApplicationDisruptionBudgetResolver) Sync(ctx context.Context) error {
-	node_names, err := r.ResolveNodes(ctx)
+	node_names, err := r.GetSelectedNodes(ctx)
 	if err != nil {
 		return err
 	}
@@ -160,7 +156,7 @@ func (r *ApplicationDisruptionBudgetResolver) CheckHealth(context.Context) error
 		return err
 	}
 
-	body, err := ioutil.ReadAll(resp.Body)
+	body, err := io.ReadAll(resp.Body)
 	if err != nil {
 		log.Fatalln(err)
 	}
@@ -172,14 +168,14 @@ func (r *ApplicationDisruptionBudgetResolver) CheckHealth(context.Context) error
 	}
 }
 
-func (adbr *ApplicationDisruptionBudgetResolver) ResolveNodes(ctx context.Context) (*set.Set, error) {
+func (r *ApplicationDisruptionBudgetResolver) GetSelectedNodes(ctx context.Context) (*set.Set, error) {
 	node_names := set.New()
 
-	nodes_from_pods, err := adbr.ResolveFromPodSelector(ctx)
+	nodes_from_pods, err := r.Resolver.GetNodesFromNamespacedPodSelector(ctx, r.ApplicationDisruptionBudget.Spec.PodSelector, r.ApplicationDisruptionBudget.Namespace)
 	if err != nil {
 		return node_names, err
 	}
-	nodes_from_PVCs, err := adbr.ResolveFromPVCSelector(ctx)
+	nodes_from_PVCs, err := r.Resolver.GetNodesFromNamespacedPVCSelector(ctx, r.ApplicationDisruptionBudget.Spec.PVCSelector, r.ApplicationDisruptionBudget.Namespace)
 	if err != nil {
 		return node_names, err
 	}
@@ -187,157 +183,8 @@ func (adbr *ApplicationDisruptionBudgetResolver) ResolveNodes(ctx context.Contex
 	return nodes_from_pods.Union(nodes_from_PVCs), nil
 }
 
-func (adbr *ApplicationDisruptionBudgetResolver) ResolveFromPodSelector(ctx context.Context) (*set.Set, error) {
-	node_names := set.New()
-	selector, err := metav1.LabelSelectorAsSelector(&adbr.ApplicationDisruptionBudget.Spec.PodSelector)
-	if err != nil || selector.Empty() {
-		return node_names, err
-	}
-	opts := []client.ListOption{
-		client.InNamespace(adbr.ApplicationDisruptionBudget.Namespace),
-		client.MatchingLabelsSelector{Selector: selector},
-	}
-	pods := &corev1.PodList{}
-	err = adbr.Client.List(ctx, pods, opts...)
-	if err != nil {
-		return node_names, err
-	}
-
-	for _, pod := range pods.Items {
-		node_names.Insert(pod.Spec.NodeName)
-	}
-	return node_names, nil
-}
-
-// NodeLabelSelectorAsRequirement converts a NodeSelectorRequirement to a labels.Requirement
-// I have not been able to find a function for that in Kubernetes code, if it exists please replace this
-func NodeLabelSelectorAsRequirement(expr *corev1.NodeSelectorRequirement) (*labels.Requirement, error) {
-	var op selection.Operator
-	switch expr.Operator {
-	case corev1.NodeSelectorOpIn:
-		op = selection.In
-	case corev1.NodeSelectorOpNotIn:
-		op = selection.NotIn
-	case corev1.NodeSelectorOpExists:
-		op = selection.Exists
-	case corev1.NodeSelectorOpDoesNotExist:
-		op = selection.DoesNotExist
-	default:
-		return nil, fmt.Errorf("%q is not a valid label selector operator", expr.Operator)
-	}
-	return labels.NewRequirement(expr.Key, op, append([]string(nil), expr.Values...))
-}
-
-// NodeSelectorAsSelector converts a NodeSelector to a label selector and field selector
-// I have not been able to find a function for that in Kubernetes code, if it exists please replace this
-func NodeSelectorAsSelector(ns *corev1.NodeSelector) (labels.Selector, fields.Selector, error) {
-	if ns == nil {
-		return labels.Nothing(), fields.Nothing(), nil
-	}
-
-	if len(ns.NodeSelectorTerms) == 0 {
-		return labels.Everything(), fields.Everything(), nil
-	}
-
-	labels_requirements := make([]labels.Requirement, 0, len(ns.NodeSelectorTerms))
-	fields_requirements := make([]string, 0, len(ns.NodeSelectorTerms))
-
-	for _, term := range ns.NodeSelectorTerms {
-		for _, expr := range term.MatchExpressions {
-			r, err := NodeLabelSelectorAsRequirement(&expr)
-			if err != nil {
-				return nil, nil, err
-			}
-			labels_requirements = append(labels_requirements, *r)
-		}
-
-		for _, expr := range term.MatchFields {
-			r, err := NodeLabelSelectorAsRequirement(&expr)
-			if err != nil {
-				return nil, nil, err
-			}
-			fields_requirements = append(fields_requirements, r.String())
-		}
-	}
-
-	label_selector := labels.NewSelector()
-	label_selector = label_selector.Add(labels_requirements...)
-	field_selector, err := fields.ParseSelector(strings.Join(fields_requirements, ","))
-	return label_selector, field_selector, err
-}
-
-func (adbr *ApplicationDisruptionBudgetResolver) ResolveFromPVCSelector(ctx context.Context) (*set.Set, error) {
-	node_names := set.New()
-	selector, err := metav1.LabelSelectorAsSelector(&adbr.ApplicationDisruptionBudget.Spec.PVCSelector)
-	if err != nil {
-		return node_names, err
-	}
-	opts := []client.ListOption{
-		client.InNamespace(adbr.ApplicationDisruptionBudget.Namespace),
-		client.MatchingLabelsSelector{Selector: selector},
-	}
-	PVCs := &corev1.PersistentVolumeClaimList{}
-	err = adbr.Client.List(ctx, PVCs, opts...)
-	if err != nil {
-		return node_names, err
-	}
-
-	pvs_to_fetch := []string{}
-
-	for _, pvc := range PVCs.Items {
-		pvs_to_fetch = append(pvs_to_fetch, pvc.Spec.VolumeName)
-	}
-
-	get_options := []client.GetOption{}
-	for _, pv_name := range pvs_to_fetch {
-		pv := &corev1.PersistentVolume{}
-
-		err = adbr.Client.Get(ctx, types.NamespacedName{Name: pv_name, Namespace: ""}, pv, get_options...)
-		if err != nil {
-			return node_names, err
-		}
-
-		node_selector := pv.Spec.NodeAffinity.Required
-		if node_selector == nil {
-			continue
-		}
-
-		opts := []client.ListOption{}
-		label_selector, field_selector, err := NodeSelectorAsSelector(node_selector)
-		if err != nil {
-			return node_names, err
-		}
-
-		if label_selector.Empty() && field_selector.Empty() {
-			// Ignore this PV
-			fmt.Printf("skipping %s, no affinity", pv_name)
-			continue
-		}
-
-		if !label_selector.Empty() {
-			opts = append(opts, client.MatchingLabelsSelector{Selector: label_selector})
-		}
-
-		if !field_selector.Empty() {
-			opts = append(opts, client.MatchingFieldsSelector{Selector: field_selector})
-		}
-
-		nodes := &corev1.NodeList{}
-		err = adbr.Client.List(ctx, nodes, opts...)
-		if err != nil {
-			return node_names, err
-		}
-
-		for _, node := range nodes.Items {
-			node_names.Insert(node.Name)
-		}
-	}
-
-	return node_names, nil
-}
-
-func (adbr *ApplicationDisruptionBudgetResolver) ResolveDisruption(ctx context.Context) (int, error) {
-	selected_nodes, err := adbr.ResolveNodes(ctx)
+func (r *ApplicationDisruptionBudgetResolver) ResolveDisruption(ctx context.Context) (int, error) {
+	selected_nodes, err := r.GetSelectedNodes(ctx)
 	if err != nil {
 		return 0, err
 	}
@@ -347,7 +194,7 @@ func (adbr *ApplicationDisruptionBudgetResolver) ResolveDisruption(ctx context.C
 	opts := []client.ListOption{}
 	node_disruptions := &nodedisruptionv1alpha1.NodeDisruptionList{}
 
-	err = adbr.Client.List(ctx, node_disruptions, opts...)
+	err = r.Client.List(ctx, node_disruptions, opts...)
 	if err != nil {
 		return 0, err
 	}
@@ -358,7 +205,7 @@ func (adbr *ApplicationDisruptionBudgetResolver) ResolveDisruption(ctx context.C
 		}
 		node_disruption_resolver := NodeDisruptionResolver{
 			NodeDisruption: &nd,
-			Client:         adbr.Client,
+			Client:         r.Client,
 		}
 		disruption, err := node_disruption_resolver.GetDisruption(ctx)
 		if err != nil {
